test(core): inline OCR document unit tests
